requirements.txt
fastapi
uvicorn[standard]
python-jobspy
pandas
httpx
python-dotenv
reportlab
PyPDF2
pydantic-settings